+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ## [7.8.0]
 
-- [#857](https://redmine.czechidm.com/issues/857) - Filters - added support for multiple filter parameters, `MultiValueMap` is used. 
+- [#857](https://redmine.czechidm.com/issues/857) - Filters - added support for multiple filter parameters, `MultiValueMap` is used.
+- [#274](https://redmine.czechidm.com/issues/274) - Filters for entity event processors was added. 
 
 
 ## [7.7.0]